- Fixed an warning of type casting. (bugreport:7081) - Removed an unecessary type cast before atoi. - Fixed Ray of Genesis Attack Formula. (bugreport:7053)
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@17074 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/atcommand.c
@@ -8611,7 +8611,7 @@ ACMD_FUNC(set) {
 		if( is_str )
 			set_var(sd, reg, (void*) val);
 		else
-			set_var(sd, reg, (void*)__64BPRTSIZE((int)(atoi(val))));
+			set_var(sd, reg, (void*)__64BPRTSIZE((atoi(val))));
 
 	}
